pyproject.toml
[tool.poetry]
name = "mwc-pedprog-unit02-lab04"
version = "0.1.0"
description = ""
authors = ["Chris Proctor <github.com@accounts.chrisproctor.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
questgame = "^0.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
